Refactor layer functions to templates/layers
- Moved all native layer functions and generation logic from `tensorcircuit/applications/layers.py` to `tensorcircuit/templates/layers.py`.
- Refactored `_resolve` in the new location to be backend-agnostic using `tc.backend`.
- Simplified `tensorcircuit/applications/layers.py` to a deprecated skeleton while maintaining its internal Cirq-related logic.
- Updated `tensorcircuit/templates/__init__.py` to export the new `layers` module.
- Fixed a bug in `bitfliplayer_mc`.
