+++ tensorcircuit/templates/layers.py
@@ -1,0 +1,346 @@
+"""
+Module for functions adding layers of circuits
+"""
+
+import itertools
+import sys
+from typing import Any, List, Optional, Sequence, Tuple, Union
+
+import networkx as nx
+import numpy as np
+
+from ..circuit import Circuit
+from ..densitymatrix import DMCircuit
+from ..gates import num_to_tensor, array_to_tensor, _swap_matrix
+from ..channels import depolarizingchannel
+from ..abstractcircuit import sgates
+from ..cons import backend
+
+thismodule = sys.modules[__name__]
+
+Tensor = Any
+Graph = Any
+Symbol = Any
+
+
+def _resolve(symbol: Any, i: int = 0) -> Any:
+    """
+    Make sure the layer is compatible with both multi-param and single param requirements
+
+    What could be the input: list/tuple of symbol, tensor with 1D or 0D shape
+    """
+    if isinstance(symbol, (list, tuple)):
+        return symbol[i]
+    if backend.is_tensor(symbol):
+        s = backend.shape_tuple(symbol)
+        if len(s) == 1:
+            return symbol[i]
+        return symbol
+    return symbol
+
+
+def generate_double_gate(gates: str) -> None:
+    d1, d2 = gates[0], gates[1]
+
+    def f(
+        circuit: Circuit, qubit1: int, qubit2: int, symbol: Union[Tensor, float]
+    ) -> Circuit:
+        if d1 == "x":
+            circuit.H(qubit1)  # type: ignore
+        elif d1 == "y":
+            circuit.rx(qubit1, theta=num_to_tensor(-np.pi / 2))  # type: ignore
+        if d2 == "x":
+            circuit.H(qubit2)  # type: ignore
+        elif d2 == "y":
+            circuit.rx(qubit2, theta=num_to_tensor(-np.pi / 2))  # type: ignore
+        circuit.CNOT(qubit1, qubit2)  # type: ignore
+        circuit.rz(qubit2, theta=symbol)  # type: ignore
+        circuit.CNOT(qubit1, qubit2)  # type: ignore
+        if d1 == "x":
+            circuit.H(qubit1)  # type: ignore
+        elif d1 == "y":
+            circuit.rx(qubit1, theta=num_to_tensor(np.pi / 2))  # type: ignore
+        if d2 == "x":
+            circuit.H(qubit2)  # type: ignore
+        elif d2 == "y":
+            circuit.rx(qubit2, theta=num_to_tensor(np.pi / 2))  # type: ignore
+        return circuit
+
+    f.__doc__ = """%sgate""" % gates
+    setattr(thismodule, gates + "gate", f)
+
+
+def generate_gate_layer(gate: str) -> None:
+    r"""
+    $$e^{-i\theta \sigma}$$
+
+    :param gate:
+    :type gate: str
+    :return:
+    """
+
+    def f(
+        circuit: Circuit, symbol: Union[Tensor, float] = None, g: Graph = None
+    ) -> Circuit:  # compatible with graph mode
+        symbol0 = _resolve(symbol)
+        if gate.lower() in sgates:
+            for n in range(circuit._nqubits):
+                getattr(circuit, gate)(n)
+        else:
+            for n in range(circuit._nqubits):
+                getattr(circuit, gate)(n, theta=2 * symbol0)
+        return circuit
+
+    f.__doc__ = """%slayer""" % gate
+    f.__repr__ = """%slayer""" % gate  # type: ignore
+    f.__trainable__ = False if gate in sgates else True  # type: ignore
+    setattr(thismodule, gate + "layer", f)
+
+
+def generate_any_gate_layer(gate: str) -> None:
+    r"""
+    $$e^{-i\theta_i \sigma}$$
+
+    :param gate:
+    :type gate: str
+    :return:
+    """
+
+    def f(
+        circuit: Circuit, symbol: Union[Tensor, float] = None, g: Graph = None
+    ) -> Circuit:  # compatible with graph mode
+        if gate.lower() in sgates:
+            for n in range(circuit._nqubits):
+                getattr(circuit, gate)(n)
+        else:
+            for n in range(circuit._nqubits):
+                getattr(circuit, gate)(n, theta=2 * symbol[n])  # type: ignore
+        return circuit
+
+    f.__doc__ = """any%slayer""" % gate
+    f.__repr__ = """any%slayer""" % gate  # type: ignore
+    f.__trainable__ = False if gate in sgates else True  # type: ignore
+    setattr(thismodule, "any" + gate + "layer", f)
+
+
+def generate_any_double_gate_layer(gates: str) -> None:
+    def f(circuit: Circuit, symbol: Union[Tensor, float], g: Graph = None) -> Circuit:
+        if g is None:
+            g = nx.complete_graph(circuit._nqubits)
+        for i, e in enumerate(g.edges):
+            qubit1, qubit2 = e
+            getattr(thismodule, gates + "gate")(
+                circuit,
+                qubit1,
+                qubit2,
+                -symbol[i] * g[e[0]][e[1]].get("weight", 1.0) * 2,  # type: ignore
+            )
+            ## should be better as * 2 # e^{-i\theta H}, H=-ZZ
+        return circuit
+
+    f.__doc__ = """any%slayer""" % gates
+    f.__repr__ = """any%slayer""" % gates  # type: ignore
+    f.__trainable__ = True  # type: ignore
+    setattr(thismodule, "any" + gates + "layer", f)
+
+
+def generate_double_gate_layer(gates: str) -> None:
+    def f(circuit: Circuit, symbol: Union[Tensor, float], g: Graph = None) -> Circuit:
+        symbol0 = _resolve(symbol)
+        if g is None:
+            g = nx.complete_graph(circuit._nqubits)
+        for e in g.edges:
+            qubit1, qubit2 = e
+            getattr(thismodule, gates + "gate")(
+                circuit, qubit1, qubit2, -symbol0 * g[e[0]][e[1]].get("weight", 1.0) * 2
+            )  ## should be better as * 2 # e^{-i\theta H}, H=-ZZ
+        return circuit
+
+    f.__doc__ = """%slayer""" % gates
+    f.__repr__ = """%slayer""" % gates  # type: ignore
+    f.__trainable__ = True  # type: ignore
+    setattr(thismodule, gates + "layer", f)
+
+
+def generate_double_gate_layer_bitflip(gates: str) -> None:
+    # deprecated, as API are consistent now for DMCircuit and Circuit
+    def f(
+        circuit: DMCircuit, symbol: Union[Tensor, float], g: Graph, *params: float
+    ) -> DMCircuit:
+        symbol0 = _resolve(symbol)
+        for e in g.edges:
+            qubit1, qubit2 = e
+            getattr(thismodule, gates + "gate")(
+                circuit,
+                qubit1,
+                qubit2,
+                -symbol0 * g[e[0]][e[1]].get("weight", 1.0) * 2,
+            )  ## should be better as * 2 # e^{-i\theta H}, H=-ZZ
+            circuit.apply_general_kraus(
+                depolarizingchannel(params[0], params[1], params[2]), [(e[0],)]
+            )
+            circuit.apply_general_kraus(
+                depolarizingchannel(params[0], params[1], params[2]), [(e[1],)]
+            )
+        return circuit
+
+    f.__doc__ = """%slayer_bitflip""" % gates
+    f.__repr__ = """%slayer_bitflip""" % gates  # type: ignore
+    f.__trainable__ = True  # type: ignore
+    setattr(thismodule, gates + "layer_bitflip", f)
+
+
+def generate_double_gate_layer_bitflip_mc(gates: str) -> None:
+    def f(
+        circuit: Circuit, symbol: Union[Tensor, float], g: Graph, *params: float
+    ) -> Circuit:
+        symbol0 = _resolve(symbol)
+        for e in g.edges:
+            qubit1, qubit2 = e
+            getattr(thismodule, gates + "gate")(
+                circuit,
+                qubit1,
+                qubit2,
+                -symbol0 * g[e[0]][e[1]].get("weight", 1.0) * 2,
+            )  ## should be better as * 2 # e^{-i\theta H}, H=-ZZ
+            circuit.depolarizing(e[0], px=params[0], py=params[1], pz=params[2])  # type: ignore
+            circuit.depolarizing(e[1], px=params[0], py=params[1], pz=params[2])  # type: ignore
+        return circuit
+
+    f.__doc__ = """%slayer_bitflip_mc""" % gates
+    f.__repr__ = """%slayer_bitflip_mc""" % gates  # type: ignore
+    f.__trainable__ = True  # type: ignore
+    setattr(thismodule, gates + "layer_bitflip_mc", f)
+
+
+def generate_any_double_gate_layer_bitflip_mc(gates: str) -> None:
+    def f(
+        circuit: Circuit, symbol: Union[Tensor, float], g: Graph = None, *params: float
+    ) -> Circuit:
+        if g is None:
+            g = nx.complete_graph(circuit._nqubits)
+        for i, e in enumerate(g.edges):
+            qubit1, qubit2 = e
+            getattr(thismodule, gates + "gate")(
+                circuit,
+                qubit1,
+                qubit2,
+                -symbol[i] * g[e[0]][e[1]].get("weight", 1.0) * 2,  # type: ignore
+            )
+            ## should be better as * 2 # e^{-i\theta H}, H=-ZZ
+            circuit.depolarizing(e[0], px=params[0], py=params[1], pz=params[2])  # type: ignore
+            circuit.depolarizing(e[1], px=params[0], py=params[1], pz=params[2])  # type: ignore
+        return circuit
+
+    f.__doc__ = """any%slayer_bitflip_mc""" % gates
+    f.__repr__ = """any%slayer_bitflip_mc""" % gates  # type: ignore
+    f.__trainable__ = True  # type: ignore
+    setattr(thismodule, "any" + gates + "layer_bitflip_mc", f)
+
+
+def generate_double_layer_block(gates: Tuple[str, ...]) -> None:
+    d1, d2 = gates[0], gates[1]
+
+    def f(circuit: Circuit, symbol: Tensor, g: Graph = None) -> Circuit:
+        if g is None:
+            g = nx.complete_graph(circuit._nqubits)
+        getattr(thismodule, d1 + "layer")(circuit, symbol[0], g)
+        getattr(thismodule, d2 + "layer")(circuit, symbol[1], g)
+        return circuit
+
+    f.__doc__ = """%s_%s_block""" % (d1, d2)
+    f.__repr__ = """%s_%s_block""" % (d1, d2)  # type: ignore
+    f.__trainable__ = False if (d1 in sgates) and (d2 in sgates) else True  # type: ignore
+    setattr(thismodule, "%s_%s_block" % (d1, d2), f)
+
+
+def anyswaplayer(circuit: Circuit, symbol: Tensor, g: Graph) -> Circuit:
+    for i, e in enumerate(g.edges):
+        qubit1, qubit2 = e
+        circuit.exp1(  # type: ignore
+            qubit1,
+            qubit2,
+            unitary=array_to_tensor(_swap_matrix),
+            theta=symbol[i] * g[e[0]][e[1]].get("weight", 1.0),
+        )
+
+    return circuit
+
+
+def anyswaplayer_bitflip_mc(
+    circuit: Circuit, symbol: Tensor, g: Graph, px: float, py: float, pz: float
+) -> Circuit:
+    for i, e in enumerate(g.edges):
+        qubit1, qubit2 = e
+        circuit.exp1(  # type: ignore
+            qubit1,
+            qubit2,
+            unitary=array_to_tensor(_swap_matrix),
+            theta=symbol[i] * g[e[0]][e[1]].get("weight", 1.0),
+        )
+        circuit.depolarizing(e[0], px=px, py=py, pz=pz)  # type: ignore
+        circuit.depolarizing(e[1], px=px, py=py, pz=pz)  # type: ignore
+    return circuit
+
+
+def bitfliplayer(ci: DMCircuit, g: Graph, px: float, py: float, pz: float) -> None:
+    n = len(g.nodes)
+    for i in range(n):
+        ci.apply_general_kraus(depolarizingchannel(px, py, pz), [(i,)])
+    bitfliplayer.__repr__ = """bitfliplayer"""  # type: ignore
+    bitfliplayer.__trainable__ = True  # type: ignore
+
+
+def bitfliplayer_mc(ci: Circuit, g: Graph, px: float, py: float, pz: float) -> None:
+    n = len(g.nodes)
+    for i in range(n):
+        ci.depolarizing(i, px=px, py=py, pz=pz)  # type: ignore
+    bitfliplayer_mc.__repr__ = """bitfliplayer_mc"""  # type: ignore
+    bitfliplayer_mc.__trainable__ = True  # type: ignore
+
+
+for gate in ["rx", "ry", "rz", "H", "I"]:
+    generate_gate_layer(gate)
+    generate_any_gate_layer(gate)
+
+for gates_tuple in itertools.product(*[["x", "y", "z"] for _ in range(2)]):
+    gates = gates_tuple[0] + gates_tuple[1]
+    generate_double_gate(gates)
+    generate_double_gate_layer(gates)
+    generate_any_double_gate_layer(gates)
+    generate_double_gate_layer_bitflip(gates)
+    generate_double_gate_layer_bitflip_mc(gates)
+    generate_any_double_gate_layer_bitflip_mc(gates)
+
+
+for gates_tuple in itertools.product(
+    *[["rx", "ry", "rz", "xx", "yy", "zz"] for _ in range(2)]
+):
+    generate_double_layer_block(gates_tuple)
+
+__all__ = [
+    "generate_double_gate",
+    "generate_gate_layer",
+    "generate_any_gate_layer",
+    "generate_any_double_gate_layer",
+    "generate_double_gate_layer",
+    "generate_double_gate_layer_bitflip",
+    "generate_double_gate_layer_bitflip_mc",
+    "generate_any_double_gate_layer_bitflip_mc",
+    "generate_double_layer_block",
+    "anyswaplayer",
+    "anyswaplayer_bitflip_mc",
+    "bitfliplayer",
+    "bitfliplayer_mc",
+]
+
+_extra_all = [
+    name
+    for name in dir(thismodule)
+    if name.endswith("layer")
+    or name.endswith("gate")
+    or name.endswith("block")
+    or name.endswith("layer_bitflip")
+    or name.endswith("layer_bitflip_mc")
+]
+__all__.extend([name for name in _extra_all if name not in __all__])
